Выносит всю логику авторизации из App в SignIn
