+++ src/components/app/app.jsx
@@ -5,26 +5,21 @@ import {Switch, Route, BrowserRouter} from "react-router-dom";
 import MoviePage from "../movie-page/movie-page.jsx";
 import {connect} from "react-redux";
 import {getAllMovies, getPromoMovie} from "../../reducer/data/selectors";
-import {Operation as DataOperation} from "../../reducer/data/data";
 import SignIn from "../sign-in/sign-in.jsx";
-import {Operation as UserOperation} from "../../reducer/user/user";
-import {getLoginErrorStatus} from "../../reducer/user/selectors";
 
 class App extends PureComponent {
   constructor(props) {
     super(props);
-    // TODO: Возможно стоит вынести incorrectEmail в Redux
+
     this.state = {
       displayedMovie: -1,
-      incorrectEmail: false,
     };
 
     this._handleMovieCardClick = this._handleMovieCardClick.bind(this);
-    this._handleFormSubmit = this._handleFormSubmit.bind(this);
   }
 
   render() {
-    const {allMovies: movies, authError} = this.props;
+    const {allMovies: movies} = this.props;
     if (movies.length === 0) {
       return (<h1>Данные загружаются</h1>);
     }
@@ -40,11 +35,7 @@ class App extends PureComponent {
             />
           </Route>
           <Route exact path="/login">
-            <SignIn
-              authError={authError}
-              incorrectEmail={this.state.incorrectEmail}
-              onSubmit={this._handleFormSubmit}
-            />
+            <SignIn/>
           </Route>
         </Switch>
       </BrowserRouter>
@@ -74,38 +65,11 @@ class App extends PureComponent {
     const movieIndex = this.props.allMovies.findIndex((movie) => movie.id === movieId);
     this.setState({displayedMovie: movieIndex});
   }
-
-  _handleFormSubmit(formData) {
-    const {login} = this.props;
-    const {loginValue, passwordValue} = formData;
-
-    const emailValid = loginValue.match(/^([\w.%+-]+)@([\w-]+\.)+([\w]{2,})$/i);
-    if (emailValid) {
-      this.setState({
-        incorrectEmail: false,
-      });
-      login({
-        login: loginValue,
-        password: passwordValue,
-      });
-    } else {
-      this.setState({
-        incorrectEmail: true,
-      });
-    }
-  }
 }
 
 const mapStateToProps = (state) => ({
   allMovies: getAllMovies(state),
   promoMovie: getPromoMovie(state),
-  authError: getLoginErrorStatus(state),
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  login(authData) {
-    dispatch(UserOperation.login(authData));
-  },
 });
 
 App.propTypes = {
@@ -126,9 +90,7 @@ App.propTypes = {
         description: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
       })
   ).isRequired,
-  authError: PropTypes.bool.isRequired,
-  login: PropTypes.func.isRequired,
 };
 
 export {App};
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default connect(mapStateToProps)(App);
